+++ javaLs/src/Direction.java
@@ -1,11 +1,11 @@
-//Перечисление направлений
+//РџРµСЂРµС‡РёСЃР»РµРЅРёРµ РЅР°РїСЂР°РІР»РµРЅРёР№
 public enum Direction {
-    //Вверх
+    //Р’РІРµСЂС…
     UP,
-    //Влево
+    //Р’Р»РµРІРѕ
     LEFT,
-    //Вправо
+    //Р’РїСЂР°РІРѕ
     RIGHT,
-    //Вниз
+    //Р’РЅРёР·
     DOWN
 }
